requirements.txt
alabaster==0.7.13
alembic==1.9.3
attrs==22.2.0
azure-ai-formrecognizer==3.2.0
azure-common==1.1.28
azure-core==1.26.3
Babel==2.11.0
backoff==2.2.1
blobfile==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.7
coverage==7.1.0
cycler==0.11.0
databricks-cli==0.17.4
dill==0.3.6
docker==6.0.1
docopt==0.6.2
docutils==0.18.1
elasticsearch==7.17.9
entrypoints==0.4
farm-haystack==1.13.1
filelock==3.9.0
Flask==2.2.2
fonttools==4.38.0
gitdb==4.0.10
GitPython==0.3.6
greenlet==2.0.2
gunicorn==20.1.0
huggingface-hub==0.12.0
idna==3.4
imagesize==1.4.1
importlib-metadata==5.2.0
inflect==6.0.2
isodate==0.6.1
itsdangerous==2.1.2
jarowinkler==1.2.3
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
langdetect==1.0.9
llvmlite==0.39.1
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
mlflow==2.1.1
mmh3==3.0.0
monotonic==1.6
more-itertools==9.0.0
msrest==0.7.1
networkx==3.0
nltk==3.8.1
nose==1.3.7
num2words==0.5.12
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
packaging==22.0
pandas==1.5.3
pdf2image==1.16.2
Pillow==9.4.0
posthog==2.3.1
protobuf==3.20.2
pyarrow==10.0.1
pycryptodomex==3.17
pydantic==1.10.4
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==3.0.9
pypi-publisher==0.0.4
pyrsistent==0.19.3
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.7.1
PyYAML==6.0
quantulum3==0.8.1
querystring-parser==1.2.4
rank-bm25==0.2.2
rapidfuzz==2.7.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
scikit-learn==1.2.1
scipy==1.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
seqeval==1.2.2
shap==0.41.0
six==1.16.0
slicer==0.0.7
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==6.1.3
sphinx-rtd-theme==1.2.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.46
sqlparse==0.4.3
tabulate==0.9.0
threadpoolctl==3.1.0
tika==2.6.0
tiktoken==0.2.0
tokenizers==0.13.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
urllib3==1.26.14
websocket-client==1.5.1
Werkzeug==2.2.2
xpdf-python==0.0.10
zipp==3.12.1
